(#1826) ensure scout client compiles on windows
